refactor(upload): move to separate file and deprecate upload
